refactor(admin): 클라이언트 프로젝트에서 admin 코드 분리 및 별도 프로젝트로 이동
