+++ src/app/contact/page.tsx
@@ -4,7 +4,7 @@ import React, { useState } from "react";
 import Header from "@/components/header";
 import Footer from "@/components/footer";
 import Banner from "@/components/banner";
-import CustomDropdown from "@/components/admin/customDropdown";
+import CustomDropdown from "@/components/customDropdown";
 
 export default function ContactPage() {
   const [category, setCategory] = useState("");
