+++ src/app/myPage/myContact/page.tsx
@@ -6,7 +6,7 @@ import React, { useState } from "react";
 import { useRouter } from "next/navigation";
 import Footer from "@/components/footer";
 import Pagination from "@/components/pagination";
-import CustomDropdown from "@/components/admin/customDropdown";
+import CustomDropdown from "@/components/customDropdown";
 
 export default function MyContact() {
   const router = useRouter();
